Love Percentage Calculator
Love calculation by a love calculator is based on certain algorithm. Once two names are entered, this calculator matches the name of the first person against some love, romance and relationship related parameters. Similarly the name of your partner/lover is matched against the same parameters. The love calculator's algorithm then detects how many parameters you both have in common. It is based on this analysis that the calculator reaches the conclusion and displays love percentage or compatibility.